pyproject.toml
[tool.poetry]
name = "kani-utils"
version = "1.2.0"
description = "kani-utils"
authors = ["Shawn T. O'Neil <shawn@tislab.org>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
click = "*"
importlib-metadata = "^4.8.0"
streamlit = ">=1.30.0"
pandasql = "^0.7.3"
pandas = "^2.1.4"
python-dotenv = "^1.0.0"
pdfplumber = "^0.10.3"
tabulate = "^0.9.0"
multidict = "^6.0.5"
kani = {extras = ["openai"], version = "^1.0.2"}
nest-asyncio = "^1.6.0"
requests = "^2.32.3"
upstash-redis = "^1.2.0"
dill = "^0.3.9"
redis = "^5.2.1"

[tool.poetry.group.dev.dependencies]
pytest = {version = ">=7.1.2"}
tox = {version = ">=3.25.1"}
pre-commit = {version = ">=3.3.3"}

[tool.poetry.group.docs]
optional = true

[tool.poetry.group.docs.dependencies]
sphinx = {version = ">=6.1.3"}
sphinx-rtd-theme = {version = ">=1.0.0"}
sphinx-autodoc-typehints = {version = ">=1.2.0"}
sphinx-click = {version = ">=4.3.0"}
myst-parser = {version = ">=0.18.1"}

[tool.poetry.scripts]
kani-utils = "kani_utils.cli:main"

[tool.poetry-dynamic-versioning]
enable = true
vcs = "git"
style = "pep440"

[tool.black]
line-length = 120
target-version = ["py38", "py39", "py310"]

[tool.ruff]
extend-ignore = [
    "D211",  # `no-blank-line-before-class`
    "D212",  # `multi-line-summary-first-line`
    ]
line-length = 120

# Allow autofix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]

# Select or ignore from https://beta.ruff.rs/docs/rules/
select = [
    "B",  # bugbear
    "D",  # pydocstyle
    "E",  # pycodestyle errors
    "F",  # Pyflakes
    "I",  # isort
    "S",  # flake8-bandit
    "W",  # Warning
]

unfixable = []
target-version = "py310"

[tool.ruff.mccabe]
# Unlike Flake8, default to a complexity level of 10.
max-complexity = 10

[tool.codespell]
skip = "*.po,*.ts,.git,pyproject.toml"
count = ""
quiet-level = 3
# ignore-words-list = ""

[build-system]
requires = ["poetry-core>=1.0.0", "poetry-dynamic-versioning"]
build-backend = "poetry_dynamic_versioning.backend"
